+++ requirements.txt
@@ -1,10 +1,9 @@
 chainlit==1.1.306
-chainlit==1.1.306
 langchain==0.2.6
 langchain-core
 langchain-community
-langchain-chroma==0.1.2
-langchain-openai==0.1.17
-sentence-transformers==3.0.1
-unstructured==0.14.10
-Markdown==3.6
+langchain-chroma
+langchain-openai
+sentence-transformers
+unstructured
+Markdown
